+++ src/vpux_compiler/include/vpux/compiler/utils/platform_resources.hpp
@@ -1,5 +1,5 @@
 //
-// Copyright (C) 2024 Intel Corporation.
+// Copyright (C) 2024-2025 Intel Corporation.
 // SPDX-License-Identifier: Apache 2.0
 //
 
@@ -15,13 +15,13 @@
 namespace vpux {
 
 constexpr Byte VPUX37XX_CMX_WORKSPACE_SIZE = Byte(1936_KB);
-constexpr Byte VPUX37XX_CMX_WORKSPACE_FRAGMENTATION_AWARE_SIZE =
-        Byte(static_cast<double>(VPUX37XX_CMX_WORKSPACE_SIZE.count()) * FRAGMENTATION_AVOID_RATIO);
+constexpr Byte VPUX37XX_CMX_WORKSPACE_FRAGMENTATION_AWARE_SIZE = Byte(
+        static_cast<int64_t>(static_cast<double>(VPUX37XX_CMX_WORKSPACE_SIZE.count()) * FRAGMENTATION_AVOID_RATIO));
 
 constexpr Byte VPUX40XX_CMX_WORKSPACE_SIZE =
         Byte(1440_KB);  // Error from feasibleAllication if 1449_KB; See E62792 and E60873
-constexpr Byte VPUX40XX_CMX_WORKSPACE_FRAGMENTATION_AWARE_SIZE =
-        Byte(static_cast<double>(VPUX40XX_CMX_WORKSPACE_SIZE.count()) * FRAGMENTATION_AVOID_RATIO);
+constexpr Byte VPUX40XX_CMX_WORKSPACE_FRAGMENTATION_AWARE_SIZE = Byte(
+        static_cast<int64_t>(static_cast<double>(VPUX40XX_CMX_WORKSPACE_SIZE.count()) * FRAGMENTATION_AVOID_RATIO));
 constexpr int VPUX37XX_MAX_DPU_GROUPS = 2;
 constexpr int VPUX40XX_MAX_DPU_GROUPS = 6;
